# apps/api/openapi/ship-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ship API",
    "version": "1.0.0",
    "description": "REST API for Ship — coding-agent chat platform. SSE chat streaming and WebSocket endpoints are documented but served outside the JSON SDK client."
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local wrangler dev"
    },
    {
      "url": "https://api.ship.example.com",
      "description": "Production (replace with deploy URL)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT from Ship web login"
      },
      "serviceAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_SECRET for server-to-server calls only"
      }
    },
    "schemas": {
      "SessionId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ChatPostBody": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 100000
          },
          "mode": {
            "type": "string",
            "maxLength": 64
          },
          "model": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "content"
        ]
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ]
          },
          "content": {
            "type": "string"
          },
          "parts": {
            "nullable": true
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "createdAt"
        ]
      },
      "ChatEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "payload": {
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "timestamp"
        ]
      },
      "ChatTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "createdAt"
        ]
      },
      "GitState": {
        "type": "object",
        "properties": {
          "branchName": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "hasChanges": {
            "type": "boolean"
          },
          "dirty": {
            "type": "boolean"
          },
          "pr": {
            "type": "object",
            "properties": {
              "number": {
                "type": "number"
              },
              "url": {
                "type": "string"
              },
              "draft": {
                "type": "boolean"
              },
              "title": {
                "type": "string"
              },
              "state": {
                "type": "string"
              },
              "headSha": {
                "type": "string"
              },
              "baseBranch": {
                "type": "string"
              }
            },
            "required": [
              "number",
              "url",
              "draft"
            ]
          },
          "prUrl": {
            "type": "string"
          },
          "prStatus": {
            "type": "string"
          },
          "repoUrl": {
            "type": "string"
          },
          "baseBranch": {
            "type": "string"
          },
          "checks": {
            "type": "object",
            "properties": {
              "state": {
                "type": "string",
                "enum": [
                  "pending",
                  "success",
                  "failure",
                  "error",
                  "neutral",
                  "unknown"
                ]
              },
              "total": {
                "type": "integer"
              },
              "pending": {
                "type": "integer"
              },
              "success": {
                "type": "integer"
              },
              "failure": {
                "type": "integer"
              }
            },
            "required": [
              "state",
              "total",
              "pending",
              "success",
              "failure"
            ]
          },
          "diff": {
            "type": "object",
            "properties": {
              "patch": {
                "type": "string"
              },
              "truncated": {
                "type": "boolean"
              },
              "files": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "filename": {
                      "type": "string"
                    },
                    "oldFilename": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "added",
                        "modified",
                        "deleted",
                        "renamed",
                        "copied",
                        "changed"
                      ]
                    },
                    "additions": {
                      "type": "integer"
                    },
                    "deletions": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "filename",
                    "status",
                    "additions",
                    "deletions"
                  ]
                }
              },
              "additions": {
                "type": "integer"
              },
              "deletions": {
                "type": "integer"
              }
            },
            "required": [
              "patch",
              "files",
              "additions",
              "deletions"
            ]
          },
          "commits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "hash": {
                  "type": "string"
                },
                "shortHash": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "authorName": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string"
                },
                "authoredAt": {
                  "type": "string"
                }
              },
              "required": [
                "hash",
                "shortHash",
                "subject",
                "authorName",
                "authorEmail",
                "authoredAt"
              ]
            }
          }
        }
      },
      "PermissionReplyBody": {
        "type": "object",
        "properties": {
          "reply": {
            "type": "string",
            "enum": [
              "once",
              "always",
              "reject"
            ]
          }
        }
      },
      "QuestionReplyBody": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "response"
        ]
      },
...
    },
    "parameters": {
      "SessionId": {
        "schema": {
          "$ref": "#/components/schemas/SessionId"
        },
        "required": true,
        "name": "sessionId",
        "in": "path"
      }
    }
  },
  "paths": {
...
    "/chat/{sessionId}": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatPostBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat events",
            "x-ship-transport": "sse",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/subscribe": {
      "get": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Compatibility idle SSE",
            "x-ship-transport": "sse",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/stop": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/retry": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/pause": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/resume": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/{sessionId}/messages": {
      "get": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "before",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/events": {
      "get": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatEvent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/tasks": {
      "get": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatTask"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/git/state": {
      "get": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Git state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitState"
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/permission/{permissionId}": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "permissionId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionReplyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission resolved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/question/{questionId}": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "questionId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReplyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question answered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/{sessionId}/question/{questionId}/reject": {
      "post": {
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "questionId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Question rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}